--- cmd/cl005/zz_generated.command.go
@@ -1,73 +1,0 @@
-package cl005
-
-import (
-	"io"
-	"os"
-
-	"github.com/giantswarm/microerror"
-	"github.com/giantswarm/micrologger"
-	"github.com/spf13/cobra"
-
-	"github.com/giantswarm/awscnfm/v12/cmd/cl005/ac000"
-)
-
-const (
-	name        = "cl005"
-	description = "Conformance tests for cluster scope cl005."
-)
-
-type Config struct {
-	Logger micrologger.Logger
-	Stderr io.Writer
-	Stdout io.Writer
-}
-
-func New(config Config) (*cobra.Command, error) {
-	if config.Logger == nil {
-		return nil, microerror.Maskf(invalidConfigError, "%T.Logger must not be empty", config)
-	}
-	if config.Stderr == nil {
-		config.Stderr = os.Stderr
-	}
-	if config.Stdout == nil {
-		config.Stdout = os.Stdout
-	}
-
-	var err error
-
-	var ac000Cmd *cobra.Command
-	{
-		c := ac000.Config{
-			Logger: config.Logger,
-			Stderr: config.Stderr,
-			Stdout: config.Stdout,
-		}
-
-		ac000Cmd, err = ac000.New(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
-	f := &flag{}
-
-	r := &runner{
-		flag:   f,
-		logger: config.Logger,
-		stderr: config.Stderr,
-		stdout: config.Stdout,
-	}
-
-	c := &cobra.Command{
-		Use:   name,
-		Short: description,
-		Long:  description,
-		RunE:  r.Run,
-	}
-
-	f.Init(c)
-
-	c.AddCommand(ac000Cmd)
-
-	return c, nil
-}
