--- cmd/cl005/zz_generated.runner.go
@@ -1,42 +1,0 @@
-package cl005
-
-import (
-	"context"
-	"io"
-
-	"github.com/giantswarm/microerror"
-	"github.com/giantswarm/micrologger"
-	"github.com/spf13/cobra"
-)
-
-type runner struct {
-	flag   *flag
-	logger micrologger.Logger
-	stdout io.Writer
-	stderr io.Writer
-}
-
-func (r *runner) Run(cmd *cobra.Command, args []string) error {
-	ctx := context.Background()
-
-	err := r.flag.Validate()
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	err = r.run(ctx, cmd, args)
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	return nil
-}
-
-func (r *runner) run(ctx context.Context, cmd *cobra.Command, args []string) error {
-	err := cmd.Help()
-	if err != nil {
-		return microerror.Mask(err)
-	}
-
-	return nil
-}
