+++ cmd/command.go
@@ -11,7 +11,6 @@ import (
 	"github.com/giantswarm/awscnfm/v12/cmd/action"
 	"github.com/giantswarm/awscnfm/v12/cmd/cl003"
 	"github.com/giantswarm/awscnfm/v12/cmd/cl004"
-	"github.com/giantswarm/awscnfm/v12/cmd/cl005"
 	"github.com/giantswarm/awscnfm/v12/cmd/completion"
 	"github.com/giantswarm/awscnfm/v12/cmd/plan"
 	"github.com/giantswarm/awscnfm/v12/cmd/version"
@@ -107,20 +106,6 @@ func New(config Config) (*cobra.Command, error) {
 		}
 	}
 
-	var cl005Cmd *cobra.Command
-	{
-		c := cl005.Config{
-			Logger: config.Logger,
-			Stderr: config.Stderr,
-			Stdout: config.Stdout,
-		}
-
-		cl005Cmd, err = cl005.New(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
 	var completionCmd *cobra.Command
 	{
 		c := completion.Config{
@@ -169,7 +154,6 @@ func New(config Config) (*cobra.Command, error) {
 	m.AddCommand(actionCmd)
 	m.AddCommand(cl003Cmd)
 	m.AddCommand(cl004Cmd)
-	m.AddCommand(cl005Cmd)
 	m.AddCommand(completionCmd)
 	m.AddCommand(planCmd)
 	m.AddCommand(versionCmd)
